+++ R/remove_ambiguous.R
@@ -20,21 +20,28 @@ ambiguous_patterns <- function(unknown = TRUE, uncultured = TRUE,
   # Add patterns for unknown taxa
   if (unknown) {
     output <- c(output, 
-                paste0("unknown", name_regex, "+"),
-                paste0("Unknown", name_regex, "+"),
-                paste0("UNKNOWN", name_regex, "+"),
-                paste0("unidentified", name_regex, "+"),
-                paste0("Unidentified", name_regex, "+"),
-                paste0("UNIDENTIFIED", name_regex, "+")
+                paste0(name_regex, "*", "unknown", name_regex, "*"),
+                paste0(name_regex, "*", "Unknown", name_regex, "*"),
+                paste0(name_regex, "*", "UNKNOWN", name_regex, "*"),
+                paste0(name_regex, "*", "unidentified", name_regex, "*"),
+                paste0(name_regex, "*", "Unidentified", name_regex, "*"),
+                paste0(name_regex, "*", "UNIDENTIFIED", name_regex, "*"),
+                paste0(name_regex, "*", "Incertae Sedis", name_regex, "*"),
+                paste0(name_regex, "*", "incertae sedis", name_regex, "*"),
+                paste0(name_regex, "*", "INCERTAE SEDIS", name_regex, "*")
     )
   }
   
-  # Add patterns for uncultured taxa
+  # Add patterns for uncultured taxa Incertae Sedis
   if (uncultured) {
     output <- c(output, 
-                paste0("uncultured", name_regex, "+"),
-                paste0("Uncultured", name_regex, "+"),
-                paste0("UNCULTURED", name_regex, "+"))
+                paste0(name_regex, "*", "uncultured", name_regex, "*"),
+                paste0(name_regex, "*", "Uncultured", name_regex, "*"),
+                paste0(name_regex, "*", "UNCULTURED", name_regex, "*"), 
+                paste0(name_regex, "*", "candidatus", name_regex, "*"),
+                paste0(name_regex, "*", "Candidatus", name_regex, "*"),
+                paste0(name_regex, "*", "CANDIDATUS", name_regex, "*")
+    )
   }
   
   # Add regex code for full matches
@@ -49,24 +56,27 @@ ambiguous_patterns <- function(unknown = TRUE, uncultured = TRUE,
 
 #' Find ambiguous taxon names
 #'
-#' Find taxa with ambiguous names, such as "unknown" or "uncultured". If
-#' you encounter a taxon name that represents an ambiguous taxon that is not
+#' Find taxa with ambiguous names, such as "unknown" or "uncultured". If you
+#' encounter a taxon name that represents an ambiguous taxon that is not
 #' filtered out by this function, let us know and we will add it.
-#' 
+#'
 #' @param taxon_names A taxmap object
-#' @inheritParams ambiguous_patterns 
-#' 
+#' @inheritParams ambiguous_patterns
+#' @param ignore_case If \code{TRUE}, dont consider the case of the text when
+#'   determining a match.
+#'
 #' @return TRUE/FALSE vector corresponding to \code{taxon_names}
-#' 
+#'
 #' @export
-is_ambiguous <- function(taxon_names, unknown = TRUE, uncultured = TRUE, name_regex = ".") {
+is_ambiguous <- function(taxon_names, unknown = TRUE, uncultured = TRUE,
+                         name_regex = ".", ignore_case = TRUE) {
   # Get patterns to filter out
   patterns <- ambiguous_patterns(unknown = unknown, uncultured = uncultured,
                                  name_regex = name_regex)
   
   # Find which taxa to filter out 
   Reduce(`|`, lapply(patterns, function(x) {
-    grepl(taxon_names, pattern = x)
+    grepl(taxon_names, pattern = x, ignore.case = ignore_case)
   }))
 }
 
@@ -88,11 +98,12 @@ is_ambiguous <- function(taxon_names, unknown = TRUE, uncultured = TRUE, name_re
 #'
 #' @export
 filter_ambiguous_taxa <- function(obj, unknown = TRUE, uncultured = TRUE, name_regex = ".", 
-                                  subtaxa = FALSE, drop_obs = TRUE,
+                                  ignore_case = TRUE, subtaxa = FALSE, drop_obs = TRUE,
                                   reassign_obs = TRUE, reassign_taxa = TRUE) {
   # Identify taxa to filter out 
-  to_remove <- is_ambiguous(obj$taxon_names(), unknown = unknown,
-                            uncultured = uncultured, name_regex = name_regex)
+  to_remove <- is_ambiguous(obj$taxon_names(), unknown = unknown, 
+                            uncultured = uncultured, name_regex = name_regex,
+                            ignore_case = ignore_case)
   
   taxa::filter_taxa(obj, to_remove, invert = TRUE, 
                     subtaxa = subtaxa, drop_obs = drop_obs,