+++ NEWS.md
@@ -14,6 +14,7 @@
 
 ### Improvements
 
+* Added functions to identift and remove taxa with ambiguous names like "unknown"
 * code from [ggrepel](https://github.com/slowkow/ggrepel) package now used to avoid overlapping labels. Thanks Kamil Slowikowski!
 * New function `heat_tree_matrix` to make plotting a pairwise matrix of heat trees for comparing treatments.
 * New parser named `parse_mothur_tax_summary` for mothur *.tax.summary file made by [classify.seqs](https://www.mothur.org/wiki/Classify.seqs).